typescript: transition browser/ to ts
